backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-generativeai==0.3.2
python-dotenv==1.0.1
pydantic>=2.6.1
python-multipart==0.0.9
elevenlabs==1.50.7
paho-mqtt==1.6.1
sounddevice==0.4.6
soundfile==0.12.1
scipy==1.12.0
numpy==1.24.3
requests==2.31.0
modal==0.55.4